Refactor folder structure and change short template name to consolego
