+++ src/Console/README.md
@@ -1,4 +1,4 @@
-# ConGo
+# ConsoleGo
 
 Created from [Janda.Go](https://github.com/Jandini/Janda.Go)
