+++ src/Console/src/ConsoleGo/ConsoleGo.csproj
