+++ src/Console/src/ConsoleGo/IMain.cs
@@ -1,4 +1,4 @@
-﻿namespace ConGo
+﻿namespace ConsoleGo
 {
     internal interface IMain
     {
