+++ src/Console/src/ConsoleGo/Settings.cs
@@ -1,4 +1,4 @@
-﻿namespace ConGo
+﻿namespace ConsoleGo
 {
     internal class Settings
     {
