+++ src/Console/src/ConsoleGo/appsettings.json
